+++ assets/src/components/component/tree/ComponentTreeNode.tsx
@@ -1,6 +1,5 @@
 import {
   ArrowTopRightIcon,
-  Card,
   IconFrame,
   Modal,
   Tooltip,
@@ -8,7 +7,7 @@ import {
 import { PipelineStageEdgeFragment } from 'generated/graphql'
 import { ComponentProps, useState } from 'react'
 import { type NodeProps, Position } from 'reactflow'
-import styled, { useTheme } from 'styled-components'
+import { useTheme } from 'styled-components'
 import isEmpty from 'lodash/isEmpty'
 
 import { useNodeEdges } from 'components/hooks/reactFlowHooks'
@@ -26,49 +25,14 @@ import {
 import { CLUSTER_PARAM_ID } from 'routes/cdRoutesConsts'
 
 import { RawYaml } from '../ComponentRaw'
-import { NodeHandle } from '../../utils/reactflow/handles'
-
-const ComponentTreeNodeSC = styled(Card)(({ theme }) => ({
-  '&&': {
-    display: 'flex',
-    alignItems: 'center',
-    overflow: 'hidden',
-    padding: `${theme.spacing.xsmall}px ${theme.spacing.medium}px`,
-    gap: theme.spacing.medium,
-    width: 240,
-  },
-  '.content': {
-    display: 'flex',
-    flexDirection: 'column',
-    alignItems: 'baseline',
-    columnGap: theme.spacing.small,
-    rowGap: theme.spacing.xxxsmall,
-    flexShrink: 1,
-    flexGrow: 1,
-    overflow: 'hidden',
-    '.name': {
-      ...theme.partials.text.body2Bold,
-      maxWidth: '100%',
-      ...TRUNCATE,
-      flexShrink: 1,
-    },
-    '.kind': {
-      ...theme.partials.text.caption,
-      maxWidth: '100%',
-      ...TRUNCATE,
-      color: theme.colors['text-xlight'],
-      marginRight: theme.spacing.xsmall,
-      flexShrink: 1,
-      flexGrow: 1,
-    },
-  },
-}))
+import { NodeBaseCard, NodeHandle } from '../../utils/reactflow/nodes'
 
 export function ComponentTreeNode({
   id,
   data,
   ...props
-}: NodeProps<TreeNodeMeta> & ComponentProps<typeof ComponentTreeNodeSC>) {
+}: NodeProps<TreeNodeMeta> & ComponentProps<typeof NodeBaseCard>) {
+  const theme = useTheme()
   const clusterId = useParams()[CLUSTER_PARAM_ID]
   const [open, setOpen] = useState(false)
   const { incomers, outgoers } = useNodeEdges(id)
@@ -91,8 +55,21 @@ export function ComponentTreeNode({
       : getResourceDetailsAbsPath(clusterId, kind, name, namespace)
 
   return (
-    <ComponentTreeNodeSC
+    <NodeBaseCard
       {...props}
+      css={{
+        alignItems: 'center',
+        overflow: 'hidden',
+        padding: `${theme.spacing.xsmall}px ${theme.spacing.medium}px`,
+        gap: theme.spacing.medium,
+        width: 240,
+        ':hover': {
+          backgroundColor:
+            theme.mode === 'light'
+              ? theme.colors['fill-two-hover']
+              : theme.colors['fill-zero-hover'],
+        },
+      }}
       clickable={clickable}
       onClick={
         !clickable
@@ -114,9 +91,27 @@ export function ComponentTreeNode({
         kind={data.kind}
         size={16}
       />
-      <div className="content">
+      <div
+        css={{
+          display: 'flex',
+          flexDirection: 'column',
+          alignItems: 'baseline',
+          columnGap: theme.spacing.small,
+          rowGap: theme.spacing.xxxsmall,
+          flexShrink: 1,
+          flexGrow: 1,
+          overflow: 'hidden',
+        }}
+      >
         {name && (
-          <p className="name">
+          <p
+            css={{
+              ...theme.partials.text.body2Bold,
+              ...TRUNCATE,
+              maxWidth: '100%',
+              flexShrink: 1,
+            }}
+          >
             <Tooltip
               label={name}
               placement="bottom"
@@ -126,7 +121,17 @@ export function ComponentTreeNode({
           </p>
         )}
         {kind && (
-          <p className="kind">
+          <p
+            css={{
+              ...theme.partials.text.caption,
+              maxWidth: '100%',
+              ...TRUNCATE,
+              color: theme.colors['text-xlight'],
+              marginRight: theme.spacing.xsmall,
+              flexShrink: 1,
+              flexGrow: 1,
+            }}
+          >
             <Tooltip label={kind}>
               <span>{kind}</span>
             </Tooltip>
@@ -154,7 +159,7 @@ export function ComponentTreeNode({
       <ModalMountTransition open={open}>
         <DetailsModal {...{ open, data, onClose: () => setOpen(false) }} />
       </ModalMountTransition>
-    </ComponentTreeNodeSC>
+    </NodeBaseCard>
   )
 }
